Add: Added clock entry consolidation
With this commit, I've added another optional argument to `org-clock-csv` that
allows consumers to 'consolidate' entries on the same task for a given day into
a single duration of time for that date.

This was especially useful in my use-case of doing time entry for billable time
out of Org.